trouble with timed lift actor:
how do we decide which message to drop and which message to keep for future generations of messageHanler?

we may need a finite set of handler states with immutable, truly partial isDefined rules and a central catchall in TimedLiftActor 
(turing it into a FSM-like contraption) - still it would be possible to drown in messages for states that never happen

PS: what about giving up on relying heavily on mutable handler actors?